// apps/client/src/app/core/api/market/marketboard-price.ts
export interface MarketboardPrice {
    Added: number;
    CreatorSignatureID: string;
    CraftSignature: string;
    ID: string;
    IsCrafted: boolean;
    IsHQ: boolean;
    Materia: any[];
    PricePerUnit: number;
    PriceTotal: number;
    Quantity: number;
    RetainerID: string;
    RetainerName: string;
    StainID: number;
    TownID: number;
}